requirements.txt
aiogram==3.19.0
python-dotenv==1.0.1
pyrogram==2.0.106
tgcrypto==1.2.5